fixup: better modules
