+++ crates/game/src/components/mod.rs
@@ -1,0 +1,6 @@
+mod dino;
+mod ground;
+mod tree;
+pub use dino::Dino;
+pub use ground::Ground;
+pub use tree::Tree;
