+++ crates/game/src/game_control.rs
@@ -43,11 +43,10 @@ pub fn screen_changes(
                 time.unpause();
                 next_screen.set(GameScreen::PlayScreen);
             }
-        } else if !window.focused && !time.is_paused()
-            && *cur_screen == GameScreen::PlayScreen {
-                time.pause();
-                next_screen.set(GameScreen::UnfocusedPauseScreen);
-            };
+        } else if !window.focused && !time.is_paused() && *cur_screen == GameScreen::PlayScreen {
+            time.pause();
+            next_screen.set(GameScreen::UnfocusedPauseScreen);
+        };
 
         if window.focused && keyboard.just_released(KeyCode::Escape) {
             if *cur_screen == GameScreen::ManuallyPauseScreen {
